% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallBash.R
\name{callSubmitter}
\alias{callSubmitter}
\title{callSubmitter}
\usage{
callSubmitter(input.bam.file.path, aspera.openssh.file)
}
\arguments{
\item{input.bam.file.path}{}

\item{aspera.openssh.file}{}
}
\description{
callSubmitter
}
\examples{

input.bam.file.path <- "/media/H_driver/2016/Submission2SRA/Guoyan_Nimer_mouse_RNA-Seq_BAM/"
aspera.openssh.file <- "/media/H_driver/2016/Submission2SRA/Christopher Williams_mouse_RNA-Seq_BAM/aspera.openssh"

callSubmitter(input.bam.file.path,aspera.openssh.file)


}
